requirements.txt
dash==2.16.1
Keras==2.4.3
lime==0.2.0.1
matplotlib==3.3.1
numpy==1.18.5
pandas==1.1.3
Pillow==8.0.0
Pillow==10.2.0
Requests==2.31.0
skimage==0.0
tensorflow==2.3.1
